fix app

+++ src/Fluxus.Core/Models/Invoice.cs
@@ -8,12 +8,12 @@ public class Invoice
     public long Id { get; private set; }
     public string? Description { get; private set; }
     public DateTime IssueDate { get; private set; }
-    public decimal SubtotalService { get; private set; }
-    public decimal SubtotalMileageAllowance { get; private set; }
-    public decimal Total { get; private set; }
+    public double SubtotalService { get; private set; }
+    public double SubtotalMileageAllowance { get; private set; }
+    public double Total { get; private set; }
 
 
-    public Invoice(long id, string? description, DateTime issueDate, decimal subtotalService, decimal subtotalMileageAllowance, decimal total)
+    public Invoice(long id, string? description, DateTime issueDate, double subtotalService, double subtotalMileageAllowance, double total)
     {
         Id = id;
         Description = description;
