+++ src/Fluxus.Core/Models/Service.cs
@@ -7,11 +7,11 @@ public class Service
     public long Id { get; set; }
     public string Tag { get; set; }
     public string? Description { get; set; }
-    public decimal ServiceAmount { get; set; }
-    public decimal MileageAllowance { get; set; }
+    public double ServiceAmount { get; set; }
+    public double MileageAllowance { get; set; }
 
 
-    public Service(long id, string tag, string description, decimal serviceAmount, decimal mileageAllowance)
+    public Service(long id, string tag, string description, double serviceAmount, double mileageAllowance)
     {
         Id = id;
         Tag = tag;
